checkUserPermission

+++ routes/user.router.js
@@ -1,46 +1,62 @@
 const router = require('express').Router();
 
-const { userValidator } = require('../validators');
-const { userController } = require('../controllers');
-const {
-    userMiddleware: {
-        checkUniqueEmail,
-        validateUserDataByDynamicParam,
-        getUserByDynamicParam
-    }
-} = require('../middlewares');
 const { middlewareVars } = require('../config');
+const { userController } = require('../controllers');
+const { User } = require('../dataBase');
+const {
+    authMiddleware,
+    generalMiddleware: {
+        validateDataByDynamicParam,
+        getItemByDynamicParam,
+        throwIfItemExist
+    },
+    userMiddleware
+} = require('../middlewares');
+const { userValidator } = require('../validators');
 
 router.post(
     '/',
-    validateUserDataByDynamicParam(userValidator.createUserValidator),
-    checkUniqueEmail,
+    validateDataByDynamicParam(userValidator.createUserValidator),
+    getItemByDynamicParam(User, middlewareVars.email),
+    throwIfItemExist(),
     userController.create
 );
 router.get(
     '/',
-    validateUserDataByDynamicParam(userValidator.getUsersValidator, middlewareVars.query),
+    validateDataByDynamicParam(userValidator.getUsersValidator, middlewareVars.query),
     userController.getAllOrByQuery
 );
 
 router.get(
     '/:user_id',
-    validateUserDataByDynamicParam(userValidator.userIdValidator, middlewareVars.params),
-    getUserByDynamicParam(middlewareVars.user_id, middlewareVars.params, middlewareVars.id),
+    validateDataByDynamicParam(userValidator.userIdValidator, middlewareVars.params),
+    getItemByDynamicParam(User, middlewareVars.user_id, middlewareVars.params, middlewareVars.id),
+    throwIfItemExist(false),
     userController.getOneById
 );
 router.patch(
     '/:user_id',
-    validateUserDataByDynamicParam(userValidator.userIdValidator, middlewareVars.params),
-    validateUserDataByDynamicParam(userValidator.updateUserValidator),
-    getUserByDynamicParam(middlewareVars.user_id, middlewareVars.params, middlewareVars.id),
-    checkUniqueEmail,
+    authMiddleware.validateAccessToken,
+    validateDataByDynamicParam(userValidator.userIdValidator, middlewareVars.params),
+    validateDataByDynamicParam(userValidator.updateUserValidator),
+    getItemByDynamicParam(User, middlewareVars.user_id, middlewareVars.params, middlewareVars.id),
+    throwIfItemExist(false),
+
+    userMiddleware.checkUserPermission([]),
+
+    getItemByDynamicParam(User, middlewareVars.email),
+    throwIfItemExist(),
     userController.updateById
 );
 router.delete(
     '/:user_id',
-    validateUserDataByDynamicParam(userValidator.userIdValidator, middlewareVars.params),
-    getUserByDynamicParam(middlewareVars.user_id, middlewareVars.params, middlewareVars.id),
+    authMiddleware.validateAccessToken,
+    validateDataByDynamicParam(userValidator.userIdValidator, middlewareVars.params),
+    getItemByDynamicParam(User, middlewareVars.user_id, middlewareVars.params, middlewareVars.id),
+    throwIfItemExist(false),
+
+    userMiddleware.checkUserPermission(['admin']),
+
     userController.deleteById
 );
